+++ DesignToolbox/DesignToolbox/Pages/Components/ComponentsPage.swift
@@ -17,6 +17,7 @@ struct ComponentsPage: View {
 
     var componentElements: [DesignToolboxElement] = [
         BadgeElement(),
+        BulletListElement(),
         ButtonElement(),
         ColoredSurfaceElement(),
         CheckboxElements(),