+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename of *Orange Business Tools*  theme to *Orange Compact* (Orange-OpenSource/ouds-ios#1292)
 
+## Added
+
+- [Library] Bullet List component (Orange-OpenSource/ouds-ios#513)
 
 ## [1.1.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.0.0...1.1.0) - 2026-01-27